% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcpr.R
\docType{package}
\name{dlcpr}
\alias{dlcpr}
\title{dlcpr}
\description{
Motion tracking and pose estimation are useful tools for various fields including psychology, neuroscience, and animal behaviour. DeepLabCut is a open source software based on deep learning, which allows multi body parts tracking on Python.
However, ones need pre-process flows when analyzing tracking data, and it often takes.
This package is for process csv files from DeepLabCut and combine to help your later analysis.
}
