% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dlc.r
\name{read.dlc}
\alias{read.dlc}
\title{This is modified function of fread from data.table package by adjusting to the format of DeepLabCut files}
\usage{
read.dlc(fileName, direc = NULL, fps = NULL, spline = TRUE, threshold = 0.9)
}
\arguments{
\item{[direc]}{optional: directory of csv files}

\item{[fps]}{optional: camera fps}

\item{[spline]}{if TRUE , drop frames of low likelihood frames and interpolate with smooth spline function. Since peformance of extrapolation is serverely bad, initial and last dropped frames would be NA. Default is TRUE.}

\item{[threshold]}{threshold for dropping frames if spine = TRUE. Default is .9}
}
\value{
data frame of trajectories
}
\description{
This is modified function of fread from data.table package by adjusting to the format of DeepLabCut files
}
