% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.lowlikli.r
\name{spline.lowlikli}
\alias{spline.lowlikli}
\title{Smooth spline function}
\usage{
spline.lowlikli(d, threshold = 0.9)
}
\arguments{
\item{fileName:}{charcters of a file name}
}
\value{
data frame of trajectories
}
\description{
Smooth spline function
}
